Modify: rename Component folder
